requirements.txt
absl-py==2.0.0
amqp==5.1.1
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.1.0
cachetools==5.3.1
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==15.1.0
contourpy==1.1.1
cron-descriptor==1.4.0
cryptography==41.0.4
cycler==0.11.0
daphne==4.0.0
Django==4.1.11
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-timezone-field==6.0.1
flatbuffers==23.5.26
fonttools==4.42.1
gast==0.4.0
google-auth==2.23.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.58.0
h5py==3.9.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
kombu==5.3.2
libclang==16.0.6
mailjet-rest==1.3.4
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.8.0
numpy==1.26.0
oauthlib==3.2.2
opencv-python==4.8.0.76
opt-einsum==3.3.0
packaging==23.1
Pillow==10.0.1
prompt-toolkit==3.0.39
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
pyparsing==3.1.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==4.5.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tqdm==4.66.1
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.7
Werkzeug==2.3.7
wrapt==1.15.0
zope.interface==6.1
